requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
oauthlib==3.2.2
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
tweepy==4.15.0
urllib3==2.3.0
